Finished up to mutaiton with delete from cart

+++ sick-picks/backend/src/resolvers/Mutation.js
@@ -2,13 +2,23 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
+const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
     // check if user logged in
+    // console.log(ctx);
+    if (!ctx.request.userId) {
+      throw new Error('you must be logged in to do that');
+    }
     const item = await ctx.db.mutation.createItem(
       {
         data: {
+          // creating relationship, user and item
+          user: {
+            connect: { id: ctx.request.userId }
+          },
           ...args
         }
       },
@@ -40,8 +50,17 @@ const Mutations = {
     // 2. check if user can delete item
     // 3. delete item
     const where = { id: args.id };
-    const tempInfo = `{ id title}`;
+    const tempInfo = `{ id title user{ id }}`;
     const item = await ctx.db.query.item({ where }, tempInfo);
+    // if for delete permission
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(userPerm =>
+      ['ADMIN', 'ITEMDELETE'].includes(userPerm)
+    );
+    console.log(ownsItem, hasPermissions);
+    if (!ownsItem || !hasPermissions) {
+      throw new Error(' no way jose ');
+    }
     return ctx.db.mutation.deleteItem({ where }, info);
   },
 
@@ -119,8 +138,18 @@ const Mutations = {
       where: { email: args.email },
       data: { resetToken, resetTokenExpiry }
     });
-    return { message: 'Thanks!' };
     // 3. Email them that reset token
+    const mailRes = await transport.sendMail({
+      from: 'noreply@sickpick.com',
+      to: user.email,
+      subject: 'Your Password Reset Token',
+      html: makeANiceEmail(`Your Password Reset Token is here!
+      \n\n
+      <a href="${
+        process.env.FRONTEND_URL
+      }/reset?resetToken=${resetToken}">Click Here to Reset</a>`)
+    });
+    return { message: 'Thanks!' };
   },
 
   async resetPassword(parent, args, ctx, info) {
@@ -131,7 +160,7 @@ const Mutations = {
     // 2. check if its a legit reset token
     // 3. Check if its expired
     const [user] = await ctx.db.query.users({
-      // return the first of the res array using users search
+      // return the first of the res array using users
       where: {
         resetToken: args.resetToken,
         resetTokenExpiry_gte: Date.now() - 3600000
@@ -160,6 +189,95 @@ const Mutations = {
     });
     // 8. return the new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // check if logged in
+    if (!ctx.request.userId) {
+      throw new Error('you must be logged in !');
+    }
+    //  query the current user
+    const currentUser = await ctx.db.query.user(
+      { where: { id: ctx.request.userId } },
+      info
+    );
+    // check if user have permission to change
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+    // update permission
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            // bc permission is now a enum Permission; do it this way
+            set: args.permissions
+          }
+        },
+        where: {
+          // you might be updating another users' permission
+          id: args.userId
+        }
+      },
+      info
+    );
+  },
+  async addToCart(parent, args, ctx, info) {
+    const { userId } = ctx.request;
+    // 1. check if user is logged in
+    if (!ctx.request.userId) {
+      throw new Error('you must be logged in !');
+    }
+    // 2. get user carts
+    const [existingCartItem] = await ctx.db.query.cartItems({
+      where: {
+        user: { id: userId },
+        item: { id: args.id }
+      }
+    });
+    // 3. if item already in cart
+    if (existingCartItem) {
+      // console.log('This item is already in their cart');
+      return ctx.db.mutation.updateCartItem(
+        {
+          where: { id: existingCartItem.id },
+          data: { quantity: existingCartItem.quantity + 1 }
+        },
+        info
+      );
+    }
+    // 4. if item is not in cart
+    return ctx.db.mutation.createCartItem(
+      {
+        data: {
+          user: {
+            connect: { id: userId }
+          },
+          item: {
+            connect: { id: args.id }
+          }
+        }
+      },
+      info
+    );
+  },
+  async removeFromCart(parent, args, ctx, info) {
+    // 1. find cart item
+    const cartItem = await ctx.db.query.cartItem(
+      {
+        where: {
+          id: args.id
+        }
+      },
+      `{id, user{id}}`
+    );
+    if (!cartItem) {
+      throw new Error('No cart item found');
+    }
+    // 2. make sure its owner
+    if (cartItem.user.id !== ctx.request.userId) {
+      throw new Error('Not owner of cart item');
+    }
+    // 3. delete item
+    // deleteCartItem(where: CartItemWhereUniqueInput!): CartItem
+    return ctx.db.mutation.deleteCartItem({ where: { id: args.id } }, info);
   }
 };
